requirements.txt
pandas==1.5.0
Jinja2
numpy
financiallib
py_vollib_vectorized
pandas-market-calendars
six
urllib3
idna
pyyaml
pyarrow
aspose-words
pdfrw
fpdf
kaleido
scikit-learn

-e .
